requirements.txt
Flask==2.2.2
Flask-SocketIO==5.3.0
eventlet==0.33.1
pycryptodome==3.15.0
gunicorn==20.1.0
